+++ flask_unchained/bundles/api/hooks/__init__.py
@@ -1,8 +1,2 @@
-class Store:
-    def __init__(self):
-        self.resources_by_model = {}
-
-        self.serializers = {}
-        self.create_by_model = {}
-        self.many_by_model = {}
-        self.serializers_by_model = {}
+from .register_resources_hook import RegisterResourcesHook
+from .register_serializers_hook import RegisterSerializersHook