+++ flask_unchained/bundles/sqlalchemy/hooks/__init__.py
@@ -1,6 +1,1 @@
 from .register_models_hook import RegisterModelsHook
-
-
-class Store:
-    def __init__(self):
-        self.models = {}