+++ tests/_bundles/vendor_bundle/hooks/__init__.py
@@ -1,3 +1,0 @@
-class Store:
-    def __init__(self):
-        self.some_data = {}